pkg/uroot: use unix.Uname on Darwin
Unify the Uname implementation by using golang.org/x/sys/unix, now that
it supports Uname on Darwin.
